offload more Gantt chart logic to GanttChartLogic file
